requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
beautifulsoup4==4.7.1
black==21.9b0
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
colorama==0.4.4
Django==3.2.7
gunicorn==20.1.0
idna==2.8
isort==5.9.3
loguru==0.5.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.3
regex==2021.9.30
requests==2.21.0
soupsieve==2.2.1
sqlparse==0.4.2
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.24.3
win32-setctime==1.0.3
